refactor#61 : recipe domain 분리
recipe domain을 분리했다
