+++ src/recipe/recipe.controller.ts
@@ -1,10 +1,7 @@
 import express, { Request, Response } from 'express';
-import Ingredient from '../ingredient/ingredient.entity';
-import DetailDescription from '../recipe-description/recipe-description.entity';
-import RecipeIngredient from '../recipe-ingredient/recipe-ingredient.entity';
-import Tag from '../tag/tag.entity';
-import { AbstractRecipeController, AbstractRecipeService } from './recipe';
-import { serving } from './recipe.entity';
+
+import { AbstractRecipeController } from './type/recipeController';
+import { AbstractRecipeService } from './type/recipeService';
 
 export default class RecipeController implements AbstractRecipeController {
 	private static instance: AbstractRecipeController;
@@ -42,43 +39,12 @@ export default class RecipeController implements AbstractRecipeController {
 	// authenticate, authorize middleware 필요
 	async createRecipe(req: Request, res: Response): Promise<void> {
 		try {
-			// const { userId } = req.body;
-			const userId = 1;
-			const recipeData = {
-				title: 'myRecipe1',
-				description: 'recipe discription',
-				thumbnailUrl: '양파 사진',
-				referenceUrl: 'www.naver.com',
-				serving: serving.DONTKNOW,
-				ingredients: [
-					{
-						ingredient: {
-							name: 'ingr1',
-							category: 'category1',
-						} as Ingredient,
-						amount: 1,
-						scale: 'kg',
-					} as RecipeIngredient,
-				],
-				detailDescriptions: [
-					{
-						imageDescription: 'desc1',
-						imageUrl: 'ttttt',
-						descriptionOrder: 1,
-					} as DetailDescription,
-				],
-			};
-
-			const tags = [
-				{
-					name: 'tag',
-				},
-			] as Tag[];
-
-			RecipeController.recipeService.createRecipe(userId, recipeData, tags);
+			const { userId, recipe } = req.body;
+			console.log(userId, recipe);
+			RecipeController.recipeService.createRecipe(userId, recipe);
 			res.status(200).send();
 		} catch (error) {
-			res.status(400).send();
+			res.status(400).send(error);
 		}
 	}
